chore: split color calculate function
