+++ src/components/Color/Colors.tsx
@@ -3,6 +3,8 @@ import _ from 'lodash'
 import styled from 'styled-components'
 import * as colors from '@assets/styles/color/semantics'
 
+import { rgba2hex, hexBrightness } from '@src/helpers/color'
+
 const ColorSet = {
   Common: {
     color_background: colors.$color_background,
@@ -114,41 +116,6 @@ const ColorBox = styled.div<{ color: string; isBright: boolean }>`
     color === '#ffffff' ? `1px solid #cbd0d8` : 'none'};
   color: ${({ isBright }) => (isBright ? 'black' : 'white')};
 `
-
-const hexBrightness = (color: string) => {
-  const hasOpacity = color.length == 9
-  const splitStr = color.substr(1).match(/(\S{2})/g)
-
-  const r = parseInt(splitStr[0], 16)
-  const g = parseInt(splitStr[1], 16)
-  const b = parseInt(splitStr[2], 16)
-  const a = hasOpacity ? parseInt(splitStr[3], 16) : 1
-
-  // brightness 계산 참고 (http://alienryderflex.com/hsp.html)
-  return Math.sqrt(((r * r * 299 + g * g * 587 + b * b * 114) * a) / 1000)
-}
-
-const rgba2hex = (color: string): string => {
-  const rgbaRegex = /^rgba?\((\d+),(\d+),(\d+),?([^,\s)]+)?/i
-  const colorRemoveSpace = color.replace(/\s/g, '')
-
-  if (!colorRemoveSpace.match(rgbaRegex)) {
-    return color
-  }
-
-  const rgb = color.replace(/\s/g, '').match(rgbaRegex)
-  const alpha = ((rgb && rgb[4]) || '').trim()
-  const hex = rgb
-    ? (Number(rgb[1]) | (1 << 8)).toString(16).slice(1) +
-      (Number(rgb[2]) | (1 << 8)).toString(16).slice(1) +
-      (Number(rgb[3]) | (1 << 8)).toString(16).slice(1)
-    : color
-
-  let alp = alpha === '' ? 1 : alpha
-  alp = ((Number(alp) * 255) | (1 << 8)).toString(16).slice(1)
-
-  return '#' + hex + alp
-}
 
 const ColorPalette = ({ themeName }) => (
   <WrapColorPalette className="color_palette">
